BackEnd/Imarket/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
parts==1.5.2
Pillow==9.5.0
products==1.1.2
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
utils==1.0.1
